refactor: добавление тестов, обновление ER-diagrama.
