+++ README.md
@@ -75,6 +75,56 @@ ER-диаграмма
 - внешний ключ film_id (ссылается на таблицу film) - id фильма;
 - внешний ключ genre_id (ссылается на таблицу genre) - id жанра;
 
+## film_director
+Содержит информацию о фильмах и режиссёрах.
+Таблица включает такие поля:
+- внешний ключ film_id (ссылается на таблицу film) - id фильма;
+- внешний ключ director_id (ссылается на таблицу director) - id режиссера;
+
+## director
+Содержит информацию об именах режиссеров. 
+Таблица включает такие поля:
+- внешний ключ director_id (ссылается на таблицу director) - id режиссера;
+- name - имя режиссера. 
+
+## event_feed
+Содержит информацию о событиях.
+Таблица включает такие поля: 
+- event_id - идентификатор события;
+- timestamp - время события;
+- внешний ключ user_id (ссылается на таблицу user) - id пользователя, с которым произошло событие;
+- event_type - тип события, может быть: 
+  - LIKE,
+  - REVIEW,
+  - FRIEND.
+- operation - тип операции, может быть:
+  - REMOVE,
+  - ADD,
+  - UPDATE.
+- entity_id - объект, с которым совершается событие (идентификатор друга, фильма и тп.). 
+  
+## reviews
+Содержит информацию об отзывах.
+Таблица включает такие поля:
+- review_id - идентификатор отзыва;
+- content;
+- is_positive;
+- внешний ключ user_id (ссылается на таблицу user) - id пользователя, который оставил отзыв;
+- внешний ключ film_id (ссылается на таблицу film) - id фильма;
+- useful;
+
+## review_likes
+Содержит информацию о лайках и дизлайках на отзывах.
+Таблица включает такие поля:
+- внешний ключ review_id (ссылается на таблицу reviews) - идентификатор отзыва;
+- внешний ключ user_id (ссылается на таблицу user) - id пользователя, который поставил лайк или дизлайк отзыву;
+- like_type - тип реакции: лайк или дизлайк.
+
+## like_type
+Содержит информацию о типах реакции: лайк или дизлайк.
+Таблица включает такие поля:
+- внешний ключ like_id (ссылается на таблицу review_likes) - идентификатор реакции;
+- name - название реакции.
 
 # Примеры запросов для основных операций приложения
 Запрос на получение всех пользователей
