+++ src/test/java/ru/yandex/practicum/db/FilmoRateApplicationUsersAndFilmAndFeedTests.java
@@ -1,8 +1,7 @@
-package ru.yandex.practicum.validation;
+package ru.yandex.practicum.db;
 
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -12,36 +11,31 @@ import ru.yandex.practicum.exception.UnknownDataException;
 import ru.yandex.practicum.model.event.Event;
 import ru.yandex.practicum.model.event.constants.EventType;
 import ru.yandex.practicum.model.event.constants.Operation;
-import ru.yandex.practicum.model.film.*;
+import ru.yandex.practicum.model.film.Film;
+import ru.yandex.practicum.model.film.Genre;
+import ru.yandex.practicum.model.film.MPA;
+import ru.yandex.practicum.model.film.Review;
 import ru.yandex.practicum.model.user.User;
-import ru.yandex.practicum.service.*;
+import ru.yandex.practicum.service.EventService;
+import ru.yandex.practicum.service.FilmService;
+import ru.yandex.practicum.service.ReviewService;
+import ru.yandex.practicum.service.UserService;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @AutoConfigureTestDatabase
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-class FilmoRateApplicationTests {
+class FilmoRateApplicationUsersAndFilmAndFeedTests {
     private final UserService userService;
     private final FilmService filmService;
     private final JdbcTemplate jdbcTemplate;
     private final ReviewService reviewService;
     private final EventService eventService;
-
-    private final DirectorService directorService;
 
     @BeforeEach
     void setUp() {
@@ -558,57 +552,6 @@ class FilmoRateApplicationTests {
     }
 
     @Test
-    void findLimitPopularFilmsByGenreAndYearTest() {
-        User user = userService.put(new User("alala@test.t", "lalala", "Liza", LocalDate.of(2002, 10, 7)));
-
-        Film film1 = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей", LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        film1.setGenres(List.of(filmService.getGenreById(2), filmService.getGenreById(6)));
-        filmService.update(film1);
-
-        Film film2 = filmService.put(new Film("Бегущий по лезвию", "Фильм про будущее", LocalDate.of(1982, 10, 9), 120, filmService.getCategoryById(2)));
-        film2.setGenres(List.of(filmService.getGenreById(4), filmService.getGenreById(6)));
-        filmService.update(film2);
-
-        Film film3 = filmService.put(new Film("Сплетница", "Сериал про сплетниц", LocalDate.of(2007, 10, 9), 45, filmService.getCategoryById(3)));
-        film3.setGenres(List.of(filmService.getGenreById(1)));
-        filmService.update(film3);
-
-        filmService.addLike(film2.getId(), user.getId());
-
-        List<Film> filmList = filmService.findLimitPopularFilmsByGenreAndYear(3, 4, 1982);
-
-        assertAll(
-                () -> assertEquals(film2, filmList.get(0), "Данные не верны"),
-                () -> assertEquals(1, filmList.size(), "Данные не верны")
-        );
-    }
-
-    @Test
-    void findPopularFilmsByYearAndGenre() {
-        User user = userService.put(new User("alala@test.t", "lalala", "Liza", LocalDate.of(2002, 10, 7)));
-
-        Film film1 = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей", LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        film1.setGenres(List.of(filmService.getGenreById(2), filmService.getGenreById(6)));
-        filmService.update(film1);
-
-        Film film2 = filmService.put(new Film("Бегущий по лезвию", "Фильм про будущее", LocalDate.of(1982, 10, 9), 120, filmService.getCategoryById(2)));
-        film2.setGenres(List.of(filmService.getGenreById(4), filmService.getGenreById(6)));
-        filmService.update(film2);
-
-        Film film3 = filmService.put(new Film("Сплетница", "Сериал про сплетниц", LocalDate.of(2007, 10, 9), 45, filmService.getCategoryById(3)));
-        film3.setGenres(List.of(filmService.getGenreById(1)));
-        filmService.update(film3);
-
-        filmService.addLike(film1.getId(), user.getId());
-
-        List<Film> filmList = filmService.findPopularFilmsByYearAndGenre(2005, 2);
-
-        assertAll(
-                () -> assertEquals(film1, filmList.get(0), "Данные не верны"),
-                () -> assertEquals(1, filmList.size(), "Данные не верны")
-        );
-    }
-
     void putAndGetEventTest() {
         User putUser = userService.put(new User("alala@test.t", "lalala", "alalala", LocalDate.now()));
         int userId = putUser.getId();
@@ -625,284 +568,5 @@ class FilmoRateApplicationTests {
         assertEquals(1, eventsUser.size(), "Событие не было добавлено.");
         assertEquals(actualEvent.getEventType(), event.getEventType(), "Произошло неверное сохранение данных эвента.");
         assertEquals(actualEvent.getUserId(), event.getUserId(), "Произошло неверное сохранение данных эвента.");
-    }
-
-    @Test
-    void findPopularFilmsByYear() {
-        User user = userService.put(new User("alala@test.t", "lalala", "Liza", LocalDate.of(2002, 10, 7)));
-
-        Film film1 = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей", LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        film1.setGenres(List.of(filmService.getGenreById(2), filmService.getGenreById(6)));
-        filmService.update(film1);
-
-        Film film2 = filmService.put(new Film("Бегущий по лезвию", "Фильм про будущее", LocalDate.of(1982, 10, 9), 120, filmService.getCategoryById(2)));
-        film2.setGenres(List.of(filmService.getGenreById(4), filmService.getGenreById(6)));
-        filmService.update(film2);
-
-        Film film3 = filmService.put(new Film("Сплетница", "Сериал про сплетниц", LocalDate.of(2007, 10, 9), 45, filmService.getCategoryById(3)));
-        film3.setGenres(List.of(filmService.getGenreById(1)));
-        filmService.update(film3);
-
-        filmService.addLike(film3.getId(), user.getId());
-
-        List<Film> filmList = filmService.findPopularFilmsByYear(2007);
-
-        assertAll(
-                () -> assertEquals(film3, filmList.get(0), "Данные не верны"),
-                () -> assertEquals(1, filmList.size(), "Данные не верны")
-        );
-    }
-
-    @Test
-    void findPopularFilmsByGenre() {
-        User user = userService.put(new User("alala@test.t", "lalala", "Liza", LocalDate.of(2002, 10, 7)));
-
-        Film film1 = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей", LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        film1.setGenres(List.of(filmService.getGenreById(2), filmService.getGenreById(6)));
-        filmService.update(film1);
-
-        Film film2 = filmService.put(new Film("Бегущий по лезвию", "Фильм про будущее", LocalDate.of(1982, 10, 9), 120, filmService.getCategoryById(2)));
-        film2.setGenres(List.of(filmService.getGenreById(4), filmService.getGenreById(6)));
-        filmService.update(film2);
-
-        Film film3 = filmService.put(new Film("Сплетница", "Сериал про сплетниц", LocalDate.of(2007, 10, 9), 45, filmService.getCategoryById(3)));
-        film3.setGenres(List.of(filmService.getGenreById(1)));
-        filmService.update(film3);
-
-        filmService.addLike(film1.getId(), user.getId());
-        filmService.addLike(film2.getId(), user.getId());
-
-        List<Film> filmList = filmService.findPopularFilmsByGenre(6);
-
-        assertAll(
-                () -> assertEquals(film1, filmList.get(0), "Данные не верны"),
-                () -> assertEquals(film2, filmList.get(1), "Данные не верны"),
-                () -> assertEquals(2, filmList.size(), "Данные не верны")
-        );
-    }
-
-    @Test
-    @DisplayName("Получение списка общих фильмов для двух пользователей")
-    void getCommonFilmsTest() {
-        Film film = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей",
-                LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-
-        User user1 = userService.put(new User("mail@mail.ru", "dolore", "Nick Name", LocalDate.now()));
-        User user2 = userService.put(new User("friend@mail.ru", "friend", "friend adipisicing", LocalDate.now()));
-
-        filmService.addLike(film.getId(), user1.getId());
-        filmService.addLike(film.getId(), user2.getId());
-
-        List<Film> actualResult = filmService.getCommonFilms(user1.getId(), user2.getId());
-        assertAll(
-                () -> assertThat(actualResult.size())
-                        .as("Длина списка общих фильмов не соответствует ожидаемой!")
-                        .isOne(),
-                () -> assertThat(actualResult.get(0))
-                        .as("Список общих фильмов не соответствует ожидаемому!")
-                        .usingRecursiveComparison()
-                        .ignoringFields("genres", "userLikes")
-                        .isEqualTo(film)
-        );
-    }
-
-    @Test
-    @DisplayName("Получение списка общих фильмов для двух пользователей, у которых их нет")
-    void getNoCommonFilmsTest() {
-        Film film = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей",
-                LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-
-        User user1 = userService.put(new User("mail@mail.ru", "dolore", "Nick Name", LocalDate.now()));
-        User user2 = userService.put(new User("friend@mail.ru", "friend", "friend adipisicing", LocalDate.now()));
-
-        filmService.addLike(film.getId(), user1.getId());
-
-        List<Film> actualResult = filmService.getCommonFilms(user1.getId(), user2.getId());
-        assertThat(actualResult.size())
-                .as("Длина списка общих фильмов не соответствует ожидаемой!")
-                .isZero();
-    }
-
-    @Test
-    @DisplayName("Получение списка общих фильмов для двух пользователей в случае, если один из них не существут")
-    void getCommonFilmsForUndefinedUser() {
-        User user1 = userService.put(new User("mail@mail.ru", "dolore", "Nick Name", LocalDate.now()));
-
-        assertThatExceptionOfType(UnknownDataException.class)
-                .isThrownBy(() -> filmService.getCommonFilms(user1.getId(), user1.getId() + 1))
-                .withMessage("user с id = %s не найден.", user1.getId() + 1);
-    }
-
-    @Test
-    @DisplayName("Получение рекомендаций для пользователя")
-    void getRecommendationsTest() {
-        Film film1 = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей",
-                LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        Film film2 = filmService.put(new Film("Бегущий по лезвию", "Фильм про будущее",
-                LocalDate.of(1998, 10, 9), 120, filmService.getCategoryById(2)));
-
-        User user1 = userService.put(new User("mail@mail.ru", "dolore", "Nick Name", LocalDate.now()));
-        User user2 = userService.put(new User("friend@mail.ru", "friend", "adipisicing", LocalDate.now()));
-
-        filmService.addLike(film1.getId(), user1.getId());
-        filmService.addLike(film1.getId(), user2.getId());
-        filmService.addLike(film2.getId(), user1.getId());
-
-        List<Film> actualResult = userService.getRecommendations(user2.getId());
-        assertAll(
-                () -> assertThat(actualResult.size())
-                        .as("Длина списка рекомендаций не соответствует ожидаемой!")
-                        .isEqualTo(1),
-                () -> assertThat(actualResult.get(0))
-                        .as("Список рекомендаций не соответствует ожидаемому!")
-                        .usingRecursiveComparison()
-                        .ignoringFields("genres", "userLikes")
-                        .isEqualTo(film2)
-        );
-    }
-
-    @Test
-    @DisplayName("Получение рекомендаций для несуществующего пользователя")
-    void getRecommendationsForUndefinedUserTest() {
-        int id = new Random().nextInt(100);
-
-        assertThatExceptionOfType(UnknownDataException.class)
-                .isThrownBy(() -> userService.getRecommendations(id))
-                .withMessage("user с id = %s не найден.", id);
-    }
-
-    @Test
-    @DisplayName("Получение рекомендаций в случае отсутствия лайков у пользователя")
-    void getRecommendationsIfNoLikesTest() {
-        User user = userService.put(new User("mail@mail.ru", "dolore", "Nick Name", LocalDate.now()));
-
-        assertThat(userService.getRecommendations(user.getId()).size())
-                .as("Длина списка рекомендаций не соответствует ожидаемой!")
-                .isZero();
-    }
-
-    @Test
-    @DisplayName("Получение рекомендаций в случае, если у наиболее похожего пользователя такой же список лайков")
-    void getRecommendationsIfSameLikesTest() {
-        Film film = filmService.put(new Film("Во все тяжкие", "Сериал про двух друзей",
-                LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-
-        User user1 = userService.put(new User("mail@mail.ru", "dolore", "Nick Name", LocalDate.now()));
-        User user2 = userService.put(new User("friend@mail.ru", "friend", "friend", LocalDate.now()));
-        List<Integer> userIds = Arrays.asList(user1.getId(), user2.getId());
-
-        userIds.forEach(userId -> filmService.addLike(film.getId(), userId));
-
-        List<Film> actualResult = userService.getRecommendations(userIds.get(new Random().nextInt(userIds.size())));
-        assertThat(actualResult.size())
-                .as("Длина списка рекомендаций не соответствует ожидаемой!")
-                .isZero();
-    }
-
-    @Test
-    @DisplayName("Добавление и получение режиссера")
-    void putAndGetDirectorTest() {
-        Director director = directorService.addDirector(new Director(1, "режиссер"));
-        assertEquals(director, directorService.getDirector(director.getId()), "не верно получены данные о режиссере");
-        assertEquals(Arrays.asList(director), directorService.getAllDirectors(),
-                "неверное количество режиссеров в базе");
-    }
-
-    @Test
-    @DisplayName("Обновление данных режиссера и получение режиссера")
-    void updateAndGetDirectorTest() {
-        Director director = directorService.addDirector(new Director(1, "режиссер"));
-        director.setName("обновили");
-        directorService.updateDirector(director);
-        assertEquals(director, directorService.getDirector(director.getId()), "не верно получены данные о режиссере");
-        assertEquals(Arrays.asList(director), directorService.getAllDirectors(),
-                "неверное количество режиссеров в базе");
-    }
-
-    @Test
-    @DisplayName("Удаление режиссера")
-    void deleteDirectorTest() {
-        assertEquals(new ArrayList<>(), directorService.getAllDirectors(), "Список не пуст");
-        Director director = directorService.addDirector(new Director(1, "режиссер"));
-        assertEquals(Arrays.asList(director), directorService.getAllDirectors(),
-                "неверное количество режиссеров в базе");
-        directorService.deleteDirector(director.getId());
-        assertEquals(new ArrayList<>(), directorService.getAllDirectors(), "Режиссер не удален");
-    }
-
-    @Test
-    @DisplayName("получение, обновление, удаление режиссера с неверным id")
-    void getUpdateDeleteDirectorByWrongIdTest() {
-        assertThrows(UnknownDataException.class, () -> directorService.getDirector(666));
-        assertThrows(UnknownDataException.class, () -> directorService.updateDirector(
-                new Director(666, "666")));
-        assertThrows(UnknownDataException.class, () -> directorService.deleteDirector(666));
-    }
-
-    @Test
-    @DisplayName("Вывод всех фильмов режиссёра, отсортированных по количеству лайков")
-    void SearchFilmsByDirectorSortedLikesTest() {
-        User userPut1 = userService.put(
-                new User("alala@test.t", "lalala", "alalala", LocalDate.now()));
-        int userId1 = userPut1.getId();
-
-        Film film1 = (new Film("Бегущий по лезвию", "Сериал про двух друзей",
-                LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        Film film2 = (new Film("Бегущий по второму лезвию", "поставили треш на поток",
-                LocalDate.of(2006, 10, 9), 100, filmService.getCategoryById(1)));
-        Film film3 = (new Film("Бегущий по третему лезвию", "поставили треш на поток",
-                LocalDate.of(2007, 10, 9), 100, filmService.getCategoryById(1)));
-
-        Director director1 = directorService.addDirector(new Director(1000, "Sprielbeg"));
-        Director director2 = directorService.addDirector(new Director(1000, "Jonson"));
-
-        film1.setDirectors(Arrays.asList(director1));
-        film1 = filmService.put(film1);
-
-        filmService.addLike(film1.getId(), userId1);
-
-        film2.setDirectors(Arrays.asList(director1));
-        film2 = filmService.put(film2);
-
-        film3.setDirectors(Arrays.asList(director2));
-
-        List<Film> sortedListFilm = Arrays.asList(filmService.get(film1.getId()), filmService.get(film2.getId()));
-
-        assertEquals(sortedListFilm, filmService.getFilmDirectorSortedBy(director1.getId(), "likes"),
-                "Неправильная сортировка по лайкам");
-    }
-
-    @Test
-    @DisplayName("Вывод всех фильмов режиссёра, отсортированных по годам")
-    void SearchFilmsByDirectorSortedYearTest() {
-        Film film1 = (new Film("Бегущий по лезвию", "Сериал про двух друзей",
-                LocalDate.of(2005, 10, 9), 100, filmService.getCategoryById(1)));
-        Film film2 = (new Film("Бегущий по второму лезвию", "поставили треш на поток",
-                LocalDate.of(2006, 10, 9), 100, filmService.getCategoryById(1)));
-        Film film3 = (new Film("Бегущий по третему лезвию", "поставили треш на поток",
-                LocalDate.of(2007, 10, 9), 100, filmService.getCategoryById(1)));
-
-        Director director1 = directorService.addDirector(new Director(1000, "Sprielbeg"));
-        Director director2 = directorService.addDirector(new Director(1000, "Jonson"));
-
-        film1.setDirectors(Arrays.asList(director1));
-        film1 = filmService.put(film1);
-
-        film2.setDirectors(Arrays.asList(director1));
-        film2 = filmService.put(film2);
-
-        film3.setDirectors(Arrays.asList(director2));
-
-        List<Film> sortedListFilm = Arrays.asList(filmService.get(film1.getId()), filmService.get(film2.getId()));
-
-        assertEquals(sortedListFilm, filmService.getFilmDirectorSortedBy(director1.getId(), "year"),
-                "Неправильная сортировка по году");
-    }
-
-    @Test
-    @DisplayName("Запрос всех фильмов режиссёра, по неверному id")
-    void SearchFilmsByDirectorSortedWrongIdTest() {
-        assertThrows(UnknownDataException.class, () -> filmService.getFilmDirectorSortedBy(666 ,"year"));
-        assertThrows(UnknownDataException.class, () -> filmService.getFilmDirectorSortedBy(666 ,"likes"));
     }
 }
